==> Part 2 - Extracting genre of each title.ipynb <==
import spotipy
import json
import re
import pandas as pd
from spotipy.oauth2 import SpotifyClientCredentials

df = pd.read_csv("C:\\Users\\Rafic Srouji\\MSBA 305\\BillBoard\\top_songs_lyrics.csv",index_col=0)

df.head()

list_artists=df['artist'].unique()

list_artists


sp = spotipy.Spotify(auth_manager=SpotifyClientCredentials(client_id="704b2381723a4a5683d0eb9a12ef734f",
                                                           client_secret="aa03bf831cc14ba9a9f6cb839de968f8"))

ids=[]
for i in list_artists:
    result = sp.search(i)
    ids.append(result['tracks']['items'][0]['artists'][0]['id'])
    

result

ids

art_genres =dict()
for i in ids:
    artist= sp.artist(i)
    if len(artist['genres'])!=0:
        art_genres[artist['name']]=artist['genres'][0]
    

artist

artist=sp.artist(ids[0])

artist['name']

artist= sp.artist(ids[36])

art_genres

with open('artist_id.txt', 'w') as f:
    for item in ids:
        f.write("%s\n" % item)

for i,j in art_genres.items():
    if re.match(".*pop|neo mellow|boy band",j):
        art_genres[i] = re.sub(".*pop|neo mellow|boy band","pop",j)
    elif re.match(".*rap",j):
        art_genres[i] = re.sub(".*rap","rap",j)
    elif re.match(".*rock",j):
        art_genres[i] = re.sub(".*rock","rock",j)
    elif re.match(".*hip hop",j):
        art_genres[i] = re.sub(".*hip hop","hip hop",j)
    elif re.match(".*country",j):
        art_genres[i] = re.sub(".*country","country",j)
    elif re.match(".*r&b|.*soul|afrofuturism",j):
        art_genres[i] = re.sub(".*r&b|.*soul|afrofuturism","r&b",j)
    elif re.match(".*latin",j):
        art_genres[i] = re.sub(".*latin","latin",j)
    elif re.match("brostep|complextro",j):
        art_genres[i] = re.sub("brostep|complextro","electro house",j)
    
    
    
    

with open('artist_genres.txt', 'w') as convert_file:
     convert_file.write(json.dumps(art_genres))

result = sp.search('Shakira')

a=[]
for i in range(10):
    a.append(list_artists[i])

b = []
for i in a:
    result = sp.search(i)
    for j in range(len(result['tracks']['items'])):
        if result['tracks']['items'][j]['artists'][0] == a:
            b.append(result['tracks']['items'][j]['artists'][0]['id'])
