==> Part 3 - Extracting Music Video Statistics using YouTube API.ipynb <==
import pandas as pd
import numpy as  np
import requests
import json
from googleapiclient.discovery import build

### Verification of API

api_key = 'AIzaSyCj-PfakFVMZ6qQS1vHMJj1IXxWKz6GJKI'
youtube = build('youtube','v3',developerKey=api_key)

f = open('C:\\Users\Rafic Srouji\\MSBA 305\\BillBoard\\artist_genres.json')
music = json.load(f)
music

### Artists' channels and playlists ID

artists = {'Daniel Powter':['UCTbOjJv4X1fk-LdfHnWAK1Q','PL4EB57EC9CA75FB97'],'Sean Paul':['UCkdc7gHpavxpgGalxKbgSHA','PLzS2lGrzLLLtTtEXBv4qtHH9eZ12vP-9O'],
           'Nelly Furtado':['UC5Y-gAeGpzgzu0ObzKnhPFA','PLUjLu4WqF_k0wzmkxRPxe63qW3ULBzQbN'] ,"James Blunt" : ['UCUg8KhMXhFxJm-YQzOQ7nPw','PLGqkITntl2rEd0Br2qpax0SCq3YbjFHUC'],
          'Shakira':['UCYLNGLIzMhRTi6ZOLjAPSmw','PL4324C6BD2E9BA7F8'],'Natasha Bedingfield':['UC6LWtQeSl8ga-EYM2R9GMvg','PL0T-BryDhUwedExwfDeK9GyXT2e5kbAYU'],
          'Gnarls Barkley':['UC6qCqY_Aftl3i6jA5Ynx0Yw','PLGJ3J_3XKnLzcez36yG0hT_NSCqPiwhzR'],'Chamillionaire':['UCqqG860EkVP_XeHrae5V6ew','PLcghV3v5_oR3P-o5WNTgOaofWaqPIXnYv'],
          'Justin Timberlake':['UC-y8ci7xfsu4L3zkSuHae0A','PL72A62438A62C82E0'],'Beyonce':['UCuHzBCaKmtaLcRAOoazhCPA','PLgXJI_F2zukJ_rA0xmnzDar3d3V8aVaG_'],
          'Rihanna':['UCcgqSM4YEo5vVQpqwN-MaNw','PL-8M5stZkYjoIoESZb5jRfOs1mWHS4LPH'],'Gwen Stefani':['UCX6mObpqyrtKMdxvCV9YkGg','PLhXWBKK3te8SyuN7fU07t2z_FSTabJwQA'],
          'Fergie':['UCDAtMQlTXquEjxFHZvJ-Mxg','PLJJOnT6bAdv-0zTiflZxGJweZNHhNozvO'],'T-Pain':['UCwff7OoSHp3el6GqnSwcSdA','PL76mRAulfEfm3aiSEAgTGu7JS8DYeKXSB'],
          'Carrie Underwood':['UCBxZZfQ8R2xtk0YEU1d8l4Q','PL5FD076661D30B42F'],'Plain White Ts':['UCf7yuABFkTc4uZhjIHuuvHg','PL1B5C3D5586358146'],
          'Akon':['UC6IBMCQ6-d7p41KHxOsq4RA','PL2YL6XQrENMCd9a5zZumt9bv8pMngqw58'],'Flo Rida':['UCBRFlde39a2U4nAkmGqJwAQ','PLbo88FM1ETaXGcgMd7fzukDqyCjaZRmDT'],
          'Leona Lewis':['UCiVxAX5STcvXQN3bYUvJi9w','PL5F249CE3CA8671F0'],'Alicia Keys':['UCK5X3f0fxO4YnVKVZP8p6hg','PLn3RU2qrqR1soSnpSZALYig3kXJ29jdnh'],
          'Lil Wayne':['UCO9zJy7HWrIS3ojB4Lr7Yqw','PLasA1IRBDbhwF5y3xByfM9jaFKwc_Ciz-'],'Timbaland':['UCC3zgQZssIUMPZEYZWaqvbg','PLVfPEn0bRuCTxcVf35egAdgpdFl4lVRuo'],
          'Jordin Sparks':['UC6m1BbuIUd3srjZKTsZWO8Q','PLAW3MlcnWMRF1IYjFnCrhA4K5WsEG_ZRk'],'Sara Bareilles':['UCWPzv6y7xDSgL86LAbkdl9w','PLs2VDJTtCn3qfgVf9iv3jGGQgpYjCnhD9'],
          'Usher':['UCaNrhBiXsXIM2epDl_kEzgQ','PLJkqBXPzFZPi4EfAfDph_FF1luHuppghn'],'Chris Brown':['UCcYrdFJF7hmPXRNaWdrko4w','PL-o37YO_uwcJJY2VT8C5Mi8hLOjJX4JFr'],
          'Black Eyed Peas':['UCBFaOy1_APEXEyA6Gws_Y1g','PLdDz77lHZ3LH4L1C7CvkVJm-sWAkCCtDM'],'Lady Gaga':['UCNL1ZadSjHpjm4q9j2sVtOA','PLf5BnbCRqFj6CLy5boY1xjh9MLn24_ijo'],
          'Taylor Swift':['UCqECaJ8Gagnn7YCbPEzWH6g','PL1CbxROoA2JjCJbRwn8TeAmZiAW8jDaDH'],'Jason Mraz':['UC8ey1FYCmgHPa9f5KxWb0mw','PLDB07C42664F065C2'],
          'Kanye West':['UCs6eXM7s8Vl5WcECcRHc2qQ','PLXu_cA-xfxWudaDanb-qFI2SAMxSYQz-K'],'The All-American Rejects':['UCLcmWbIQrvGd3cGSPpwcwQQ','PLw7LuFzpErOwry0lFXQB8XxeienZbMoUA'],
          'Kesha':['UCh1bAdm-p_kpf_-a2vpCyHw','PLCuCdnh8-13e_Ul4hGO0U30BxIMCv4RaB'],'Lady Antebellum':['UCuH-CuAIQsMq49mXNLOEYnw','PLJGbU7-jQFEXSgMd94PeW5fAdGHnmmQWH'],
          'Train':['UC4-bN_6SHokeYKaM3hNeojA','PLB1c9ltaeVLbX-n9EL1bBc0VfnI1E57ab'],'Katy Perry':['UCYvmuw-JtVrTZQ-7Y4kd63Q','PLz68cPKdtIOxcO2jNfWrKN-4qY6QyDvw_'],
          'BoB':['UCU1JusGzZe0Msn79JB3g46w','PLRVGqsN6BVIQlCD9Tt7FXw0TJ8ZTR0ael'],'Eminem':['UCfM3zsQsOnfWNUppiycmBuw','PLRgSHCeagEV4KTedHndOxXLXqZr4okOCA'],
          'Taio Cruz':['UC3tg0XIy-VL4z1rVJOtllFA','PLL3loHcMUlHEboHT5aihGO2Cjrg5WZd3x'],'Adele':['UCsRM0YB_dabtEPGPTKo-gcw','PLBrbKtPlBbNdDoA6Yw1EfEk2b-PFvb6-N'],
          'LMFAO':['UCk78ZcA6kflEvBR0UrGDH0Q','PLhiQtmFA41-ZwIRUYdfKPrisGBBvm7iFx'],'Pitbull':['UCv8nzwVPQDRjkPCkEsOdEwA','PLuNzHiV4iahl263VKqhkWQy0lDGDOGABO'],
          'Bruno Mars':['UCoUM-UJ7rirJYP8CQ0EIaHA','PL2gNzJCL3m_9QZh_MFe4wWtnO3tl-bgg5'],'Nicki Minaj':['UC3jOd7GUMhpgJRBhiLzuLsg','PLN4C_R4ZZCriMa2iTVxdbuFaKveKbv1xg'],
          'Maroon 5':['UCBVjMGOIkavEAhyqpxJ73Dw','PL447CE9E513965C65'],'Jennifer Lopez':['UCr8RjWUQ_9KYcIPmWiqBroQ','PL0WnuvyfeQnHmKohW6WQseiOMPUT6z1F4'],
          'Gotye':['UCFC9LamNMmLioW643VZ40OA','PLB-AovDkjw9O2idxzbAGsQ8oXesECOG1e'],'Carly Rae Jepsen':['UC-4NAdFIP6h9BmtGUFL94Pg','PLX4FGpGGwqW8y7SYmMSi7qMBICNLmv7kd'],
          'fun':['UCuLfR3hDx9lUrVWrNVYdWhA','PLzQP6g4y4msO4nx0Mlhl3XTmynkZ6L2RU'],'Ellie Goulding':['UCt-am4x4ZyeohfgM85qFgdQ','PLjj-ow931NgE9GqmAKOJjQC2znq8yQp4c'],
          'The Wanted':['UCD0BJ0znkCCrPjIsxX8OE1Q','PL922FEB1D52C19154'],'Kelly Clarkson':['UCoRUmyNL8KEYftmxRA2n3SQ','PLdq8YZzQBHEsYXfKUM5KAmd9FVy4h2c2c'],
          'One Direction':['UCb2HGwORFBo94DmRx4oLzow','PLbatIuSsTsipxll7OrDiarXXKfCuHH_OH'],'Macklemore and Ryan Lewis':['UCXYRdIXDdeZIf816EWAr5zQ','PLpeL_bS-O35V4Q5mi7pivhQCR_KHYwrU5'],
          'Robin Thicke':['UCFHZMfLTN8Zg2sUIrLJhRQQ','PLdMaA_QzF_mGi2POUi2wi28utoDqMEPsa'],'Imagine Dragons':['UCT9zcQNlyht7fRlcjmflRSA','PL9LUD5Kp855KGgwLazfJ7ed2pgBNcutCQ'],
          'P!nk':['UCE5yTn9ljzSnC_oMp9Jnckg','PLb39i192nlZA9FydkdDvfI1LwlUWEhUsJ'],'Florida Georgia Line':['UCRlwyGpIzfhZ6Rdj8mAVyEQ','PLggPK4S5-fhhYdRe8aI0NK7JQkPek63tG'],
          'Pharrell Williams':['UCNUbNl2U6Hg8J0Zem6hzC2g','PLr6xPtmPzp4rGT5qEeNCsb1s9PiTIXbz9'],'John Legend':['UCEa-JnNdYCIFn3HMhjGEWpQ','PL5odlUFMuUzzKmkiEuqFk3YFBossORfYB'],
          'Iggy Azalea':['UCGtjnDLpKGWKU6xtjB0NeKA','PLSBZNZWQnOuiPPhJrODtMenmBYl-60DIs'],'OneRepublic':['UCi4EDAgjULwwNBHOg1aaCig','PL8F17C150B752DFD2'],
          'Jason Derulo':['UCe3yZzUwpmy2eKKmF9svX0Q','PLL8N99hsW_msFOgfNC38JXg0WC_cL8kMT'],'MAG!C':['UCZ2KbFug6U6DO02xIIu2JAA','PLHHS8kyno9PyDeCQa72PC_84rrvPItgmI'],
          'Meghan Trainor':['UCkXgEcpoTE4tHsebYBouWpA','PLuqyTcRjDR7y1D-ZHRBAwhk2JCewupc0q'],'Ariana Grande':['UC9CoOnJkIBMdeijd9qYoT_g','PLFO0esT_hEQQZVKe_D-fP1HSEgK-TTYaC'],
          'Sam Smith':['UCvpDeGlR5wLP9Z3Tb6K0Xfg','PLluhO3FgxrHtJazhI0B2oiCALruYYO8Hv'],'Mark Ronson':['UCSY5V7BJSrxNvvBF0Drslow','PLmBeBcXnjN1-pHiiMF0pTWqvHwbMjd1gQ'],
          'Ed Sheeran':['UC0C-w0YjGpqDXGB8IHb662A','PLjp0AEEJ0-fGi7bkjrGhBLUF9NMraL9cL'],'Wiz Khalifa':['UCVp3nfGRxmMadNDuVbJSk8A','PLF9D6F1E324282593'],
          'Fetty Wap':['UC_eQfAQjj6mYj92l3SBmIdg','PLZYqBBTii9A5ju0k_faG7ScQ5LTXHrDIZ'],'WALK THE MOON':['UCQXfdqRai1lm0BQ6rI5wUVg','PLW9J35yUmMBLpajzIgIGCKioHO6aC-H6J'],
          'Silento':['UCS9Lijwamx03o16qbx2gEjw','PLmbzZf7-YNpcuE3tWjR-2z7bOPnteNksP'],'The Weeknd':['UC0WP5P-ufpRfjbNrmOWwLBQ','PLVt9N9S0CzvixSs0UgEP1HVF90VEpH7D_'],
          'Justin Bieber':['UCIwFjwMjI0y7PDBVEO9-bkQ','PL7cJWYlMnhfx5_hxty6mHMiTknnqIdB3f'],'Drake':['UCByOQJjav0CUDwxCk-jVNRQ','PL4TrGu5rwzH8vb3bAMvgLLjSNPshF90tB'],
          'twenty one pilots':['UCBQZwaNPFfJ1gZ1fLZpAEGw','PL0A42C164824B74F9'],'Desiigner':['UCAfvFXvjzEupwo2NtyDhRZA','PLwoUeVM6pfHR4SnSXQEYHn_RscPvPRshX'],
          'The Chainsmokers':['UCq3Ci-h945sbEYXpVlw7rJg','PLU2N0DH_rOrz9LkwMJ63_U_Tg6wBospHi'],'Kendrick Lamar':['UC3lBXcrKFnFAFkfVk5WuKcQ','PLIen3IPVo_U1lR4Kr2CdCkWwbeI0ds572'],
          'Coldplay': ['UCDPM_n1atn2ijUwHd0NNRQw','PLsvoYlzBrLFAJd4hNQSHw1lYjDKeQB_iU'],'Migos':['UC9YcTIQuhwgoOQqYMKYqW9A','PLfQALVM3SEjIOd2SugYAmiFVnPqcjw1ux'],
          'Sam Hunt':['UCeO1KnglIIZ4IO5oLm9GM4A','PLRFelV4oiIKM7F2VzNS1ehj2IDWmTvWhQ'],'Post Malone':['UCeLHszkByNZtPKcaVXOCOQQ','PLrZOPrth3tHPYt-dRZjRZCNJWjOdnjV88'],
          'Bebe Rexha':['UC5-gWZXAQqSGVfPHkA7NRiQ','PLcOgsKxgsq0XKDQg8W-SdchlEk8jXcTi5'],'Camila Cabello':['UCio_FVgKVgqcHrRiXDpnqbw','PL8Gifh-G6-ctk_eg4gcBj69vgVffKsxP5'],
          'Cardi B':['UCxMAbVFmxKUVGAll0WVGpFw','PLe9ajjFHyFINsIXI8IPF5DjRuXMCdVld7'],'Zedd':['UCPNokRZ9hacjIQ3IQL6HNUQ','PL8zLEHzcmnHavuWAuID6knTTTS0FCc-QF'],
          'Lil Nas X':['UC_uMv3bNXwapHl8Dzf2p01Q','PLbNKjAnSNvqpYdCImQys_xHqL3897rnVl'],'Halsey':['UCOCgB3xd-B-1qAm-hR9OLrA','PL4u7-QW-Ue3p2QG-PZn1b50EBo9YzU7bL'],
          'Billie Eilish':['UCiGm_E4ZwYSHV3bcW1pnSeQ','PLsCPTY_MPoPbCftVtuvquFs9ayLDwD237'],'Marshmello':['UCEdvpU2pFRCVqU6yIPyTpMQ','PLcYK4PlHbZQvbWClI38SfOjBt3godBsY1'],
          'Khalid':['UCkntT5Je5DDopF70YUsnuEQ','PLedovvx3ekTAY0gn-cWf3SJgSUDBW2pho'],'Travis Scott':['UCtxdfwb9wfkoGocVUAJ-Bmg','PLrzEPoI2gTMA33WPq7cH8zU33i13VKtX8'],
          'Jonas Brothers':['UCFvXnyAhluG3sa6p3eOs_LA','PLISjWix87QOmCbVCUY5g-vEMSgUITz575'],'Roddy Ricch':['UChQdA1rid5kKZh6oIc6DLNg','PLoIBqO622uqi5ILoke96uXewjPf0mIh4C'],
          'Dua Lipa':['UC-J-KZfRV8c13fOCkhXdLiQ','PLNrotoZZ8BaoXT_LJuwEyESQlctWNDCwD'],'DaBaby':['UC2CCXzC56k0b8L49R_iW-Yw','PLgZ4DdsdoXP0Q8A-8hZLLmBG95G50lIMi'],
          'Harry Styles':['UCZFWPqqPkFlNwIxcpsLOwew','PL1BwKXJqW9zW2aeWGc2yjn1y9JzsCbVmD'],'Future':['UCSDvKdIQOwTfcyOimSi9oYA','PLJ6z_vXJKxu01GLF2oZnTmaNkLrcu6l8J'],
          'Maren Morris':['UCDw6bxq6sJOBxcexivmmObA','PLmnP2Opb2ewhOaJ6AvOfjiglIh_PJoyg5'],'Lewis Capaldi':['UCveFkLdSOUsGwMJEgedO9dQ','PLLj8bU170Es1sirAWDDllJgsW5CwVNYye']}

len(artists)

## Functions

channel_name = []
channelid = []
viewers = []
subscribers = []
date = []
playlist_id = []
num_videos=[]
images=[]
def get_channel_summary(dit):
    for i,j in dit.items():
        summary=youtube.channels().list(part='statistics',id=j[0]).execute()
        published=youtube.channels().list(part='snippet',id=j[0]).execute()
        channel_name.append(i)
        channelid.append(j[0])
        viewers.append(summary['items'][0]['statistics']['viewCount'])
        subscribers.append(summary['items'][0]['statistics']['subscriberCount'])
        date.append(published['items'][0]['snippet']['publishedAt'])
    
def get_playlist_data(dit):
        pl_data = dict()
        for i,j in dit.items():
                nextPage = None
                pl_data[i]=[]
            
                while 1:
                    pl =youtube.playlistItems().list(playlistId=j[1],maxResults=50,part='snippet',pageToken=nextPage).execute()
                    pl_data[i]=pl_data[i]+pl['items']
                    nextPage = pl.get('nextPageToken')
                    if nextPage is None:
                            break
        return pl_data
def get_image(dit):
    for i,j in dit.items():
        image=youtube.channels().list(part="snippet",id=j[0]).execute()
        images.append(image['items'][0]['snippet']['thumbnails']['medium']['url'])

def download_image(dit):        
    j=0
    for i in dit.keys():
        
            response = requests.get(images[j])
            file = open(i+".png", "wb")
            file.write(response.content)
            file.close()
            if j == len(dit)-1:
                break
            j=j+1  
    
def get_title_vid_id(data,list1,list2,dit,dit1):
    for i,j in data.items():
        dit[i]=[]
        dit1[i]=[]
        for m in range(0,len(data[i])):
            list1.append(data[i][m]['snippet']['title'])
            list2.append(data[i][m]['snippet']['resourceId']['videoId'])
            dit[i].append(data[i][m]['snippet']['title'])
            dit1[i].append(data[i][m]['snippet']['resourceId']['videoId'])
        
def get_artists(list1):
    for i,j in artist_songs.items():
        for x in range(len(j)):
            list1.append(i)
    return list1

published = []
likes=[]
dislikes=[]
views=[]
comment=[]
stats=[]
snippet =[]
def get_info(dataframe,list1,list2):
    for i in range(len(dataframe)):
        
        x = youtube.videos().list(part='statistics',id=dataframe.iloc[i,0]).execute()
        y = youtube.videos().list(part='snippet',id=dataframe.iloc[i,0]).execute()
        list1.append(x)
        list2.append(y)
        
def get_video_statistics(dataframe,x,y,list1,list2,list3,list4,list5):
    for i in range(len(dataframe)):
        list2.append(x[i]['items'][0]['statistics']['likeCount'])
        list3.append(x[i]['items'][0]['statistics']['dislikeCount'])
        list4.append(x[i]['items'][0]['statistics']['viewCount'])
        list5.append(x[i]['items'][0]['statistics']['commentCount'])
    for j in range(len(dataframe)):
        list1.append(y[j]['items'][0]['snippet']['publishedAt'])    
    

get_channel_summary(artists)

df = pd.DataFrame({'channelId':channelid,'channel_name':channel_name,'viewers':viewers,'subscribers':subscribers,'published':date})
df

df.to_csv("channels.csv")

d = get_playlist_data(artists)

titles=[]
vid_id=[]
artist_songs=dict()
artist_vid_id = dict()
get_title_vid_id(d,titles,vid_id,artist_songs,artist_vid_id)

titles

vid_id

artist_songs

artist_vid_id

all_artist=[]
df1= pd.DataFrame({'video_id':vid_id,'artist':get_artists(all_artist),'title':titles})

df1=df1[df1["title"].str.contains("Private video")==False]

df1

df1.to_csv("videoids.csv")

all_artist=[]
for i,j in artist_songs.items():
    for x in range(len(j)):
        all_artist.append(i)

def get_artists(list1):
    for i,j in artist_songs.items():
        for x in range(len(j)):
            list1.append(i)
    return list1

get_info(df1,stats,snippet)

stats[500]['items'][0]['statistics']['likeCount']

stats[0]['items'][0]['statistics']['likeCount']

get_video_statistics(df1,stats,snippet,published,likes,dislikes,views,comment)

likes = []
dislikes=[]
views = []
comments = []
not_sel = []
for i in range(0,len(stats)):
    if len(stats[i]['items'][0]['statistics']) == 5:
            likes.append(stats[i]['items'][0]['statistics']['likeCount'])
            dislikes.append(stats[i]['items'][0]['statistics']['dislikeCount'])
            views.append(stats[i]['items'][0]['statistics']['viewCount'])
            comments.append(stats[i]['items'][0]['statistics']['commentCount'])
    else:
        not_sel.append(stats[i]['items'][0]['id'])

for j in range(len(df1)):
    for i in range(len(not_sel)):
        if df1['video_id'][j] == not_sel[i]:
            df1.drop(index=j)

not_sel

df1.drop(df1.loc[df1['video_id']=='nWQbrYohKmE'].index, inplace=True)

df1.reset_index(inplace=True)
df1.to_csv("videoids.csv")

df1['likes']=None
df1['dislikes'] = None
df1['views'] = None
df1['comments'] = None
for i in range(len(df1)):
    df1['likes'][i]=likes[i]
    df1['dislikes'][i]=dislikes[i]
    df1['views'][i]=views[i]
    df1['comments'][i]=comments[i]
    

df1.to_csv("placeholder.csv")

comments_f = open("comments.txt", "w")
for element in views:
    comments_f.write(element + "\n")
comments_f.close()


len(likes)

len(stats)

get_image(artists)
images

download_image(artists)

df1.to_csv('YouTube.csv',index=False)
