==> Part 5 - SQL.ipynb <==
### Import MySQL Libraries

import pandas as pd
import numpy as np
import mysql.connector
from mysql.connector import Error
import datetime as dt

%store -r HOST
%store -r USER
%store -r PASS

host=HOST
user = USER
password = PASS

### Connecting to MySQL Server

def create_server_connection(host_name, user_name, user_password):
    connection = None
    try:
        connection = mysql.connector.connect(
        host=host_name,
        user=user_name,
        passwd=user_password
        )
        print("MySQL Database connection successful")
    except Error as err:
        print(f"Error: '{err}'")
    return connection

### Creating a New Databse

def create_database(connection, query):
    cursor = connection.cursor()
    try:
        cursor.execute(query)
        print("Database created successfully")
    except Error as err:
        print(f"Error: '{err}'")

### Connecting to the Database

def create_db_connection(host_name, user_name, user_password, db_name):
    connection = None
    try:
        connection = mysql.connector.connect(
        host=host_name,
        user=user_name,
        passwd=user_password,
        database=db_name
        )
        print("MySQL Database connection successful")
    except Error as err:
        print(f"Error: '{err}'")
    return connection

### Creating a Query Execution Function

def execute_query(connection, query):
    cursor = connection.cursor()
    try:
        cursor.execute(query)
        connection.commit()
        print("Query successful")
    except Error as err:
        print(f"Error: '{err}'")

### Reading Data

def read_query(connection, query):
    cursor = connection.cursor()
    result = None
    try:
        cursor.execute(query)
        result = cursor.fetchall()
        return result
    except Error as err:
        print(f"Error: '{err}'")

### Creating Records from Lists

def execute_list_query(connection, sql, val):
    cursor = connection.cursor()
    try:
        cursor.executemany(sql, val)
        connection.commit()
        print("Query successful")
    except Error as err:
        print(f"Error: '{err}'")


connection = create_server_connection(host, user, password)
create_database(connection, "CREATE DATABASE IF NOT EXISTS music;")

connection.close()

### Creating the Channel Table

connection = create_db_connection(host,user,password,'music')
create_music_table = """ CREATE TABLE channels (
                            channel_id varchar(255) PRIMARY KEY,
                            channel_name varchar(255),
                            viewers bigint,
                            subscribers bigint,
                            published date); """

execute_query(connection,create_music_table)

### Creating the Video Table

create_vid_table = """ CREATE TABLE videos (
                            video_id varchar(255) PRIMARY KEY,
                            artist varchar(255),
                            title varchar(255),
                            likes bigint,
                            dislikes bigint,
                            views bigint,
                            comments bigint,
                            channel_id varchar(255),
                            FOREIGN KEY (channel_id) REFERENCES channels(channel_id)); """

execute_query(connection,create_vid_table)

channels = pd.read_csv("channels.csv",index_col=0)

channels.head()

### Convert Published Date Column to a Date Type

channels['published']=pd.to_datetime(channels['published'])

channels['published']=pd.to_datetime(channels['published'],format="%Y-%m-%d")

channels['published']=channels['published'].dt.date

channels.head()

genres=pd.read_csv("genres.csv",index_col=0)

genres.head()

video = pd.read_csv("video.csv",index_col=0)

video

video.head()

video['channel_id']=None
for i in range(len(channels)):
    for j in range(len(video)):
        if channels['channel_name'][i] == video['artist'][j]:
            video['channel_id'][j] = channels['channelId'][i]

video.head(50)

channel_ids = []
channel_name = []
channel_views = []
channel_subscribers = []
channel_published = []
for i in range(len(channels)):
    channel_ids.append(channels['channelId'][i])
    channel_name.append(channels['channel_name'][i])
    channel_views.append(channels['viewers'][i])
    channel_subscribers.append(channels['subscribers'][i])
    channel_published.append(channels['published'][i])

### Inserting Rows into the Channel Table

sql= '''INSERT INTO channels(channel_id,channel_name,viewers,subscribers,published) VALUES (%s,%s,%s,%s,%s)'''

records = []

for i in range(len(channels)):
    records.append([channel_ids[i],channel_name[i],np.int64(channel_views[i]).item(),np.int64(channel_subscribers[i]).item(),channel_published[i]])

#Pandas stores the integer as numpy.int64 or numpy.int32 which MySQL does not support
#So the values must be converted to base int using the np.int64().item()
conenction = create_db_connection(host,user,password,'music')
execute_list_query(connection,sql,records)
connection.close()
    
    


video.drop(video.loc[video['video_id']=='#NAME?'].index,inplace=True)

video.drop_duplicates(subset=['video_id'],inplace=True)

video.reset_index(inplace=True)

video_ids = []
video_artist = []
video_title = []
video_likes = []
video_dislikes = []
video_views = []
video_comments = []
video_channel = []
for i in range(len(video)):
    video_ids.append(video['video_id'][i])
    video_artist.append(video['artist'][i])
    video_title.append(video['title'][i])
    video_likes.append(video['likes'][i])
    video_dislikes.append(video['dislikes'][i])
    video_views.append(video['views'][i])
    video_comments.append(video['comments'][i])
    video_channel.append(video['channel_id'][i])

len(video_ids)

### Inserting Rows Into the Video Table

sql= '''INSERT INTO videos(video_id,artist,title,likes,dislikes,views,comments,channel_id) VALUES (%s,%s,%s,%s,%s,%s,%s,%s)'''

records = []

for i in range(len(video)):
    records.append([video_ids[i],video_artist[i],video_title[i],np.int64(video_likes[i]).item(),np.int64(video_dislikes[i]).item(),np.int64(video_views[i]).item(),np.int64(video_comments[i]).item(),video_channel[i]])

    
conenction = create_db_connection(host,user,password,'music')
execute_list_query(connection,sql,records)
connection.close()
